+++ examples/rails_app/test/integration/trips_api_test.rb
@@ -1,8 +1,6 @@
 require 'test_helper'
 
 class TripsApiTest < ActionDispatch::IntegrationTest
-  include OpenapiFirst::Test::Methods
-
   test 'GET /trips' do
     get '/trips',
         params: { origin: 'efdbb9d1-02c2-4bc3-afb7-6788d8782b1e', destination: 'b2e783e1-c824-4d63-b37a-d8d698862f1d',