+++ examples/rails_app/test/test_helper.rb
@@ -1,20 +1,14 @@
 ENV['RAILS_ENV'] ||= 'test'
-require_relative '../config/environment'
-require 'rails/test_help'
 
 require 'openapi_first'
-OpenapiFirst::Test.register(Rails.root.join('../../spec/data/train-travel-api/openapi.yaml'))
-OpenapiFirst::Test.register(Rails.root.join('../../spec/data/attachments_openapi.yaml'), as: :attachments)
-
-OpenapiFirst::Test::Coverage.register(
-  Rails.root.join('../../spec/data/train-travel-api/openapi.yaml'),
-  Rails.root.join('../../spec/data/attachments_openapi.yaml')
-)
-OpenapiFirst::Test::Coverage.start
-
-Minitest.after_run do
-  OpenapiFirst::Test::Coverage.report
+OpenapiFirst::Test.setup do |test|
+  test.register Rails.root.join('../../spec/data/train-travel-api/openapi.yaml')
+  test.register Rails.root.join('../../spec/data/attachments_openapi.yaml'), as: :attachments
+  test.coverage_formatter_options = { verbose: true }
 end
+
+require_relative '../config/environment'
+require 'rails/test_help'
 
 module ActiveSupport
   class TestCase
@@ -25,12 +19,9 @@ module ActiveSupport
   end
 end
 
-TEST_APP = OpenapiFirst::Test.app(TrainTravel::Application, spec: Rails.root.join('../../spec/data/attachments_openapi.yaml'))
 
 module ActionDispatch
   class IntegrationTest
-    def app
-      TEST_APP
-    end
+    include OpenapiFirst::Test::Methods[TrainTravel::Application]
   end
 end